+++ src/components/ForgotPassword/ForgotPassword.js
@@ -1,7 +1,5 @@
 import React, { useState } from 'react';
-import {
-  Redirect,
-} from 'react-router-dom';
+import { useHistory } from 'react-router';
 
 import processBadResponse from '../../processBadResponse';
 import { defaultErrors } from '../../constants';
@@ -9,28 +7,23 @@ import ForgotPasswordForm from './ForgotPasswordForm';
 import useForgotPasswordMutation from './useForgotPasswordMutation';
 
 const ForgotPassword = () => {
-  const [userEmail, setUserEmail] = useState(null);
+  const history = useHistory();
   const [authFailure, setAuthFailure] = useState([]);
   const sendReminderMutation = useForgotPasswordMutation();
 
   const onSubmit = async (values) => {
-    setUserEmail(null);
     setAuthFailure([]);
     const { userInput } = values;
     const { FORGOTTEN_PASSWORD_CLIENT_ERROR } = defaultErrors;
 
     try {
       await sendReminderMutation.mutateAsync(userInput);
-      setUserEmail(userInput);
+      history.push('/check-email', { userEmail: userInput });
     } catch (error) {
       const res = await processBadResponse(undefined, error.response, FORGOTTEN_PASSWORD_CLIENT_ERROR);
       setAuthFailure(res);
     }
   };
-
-  if (userEmail) {
-    return <Redirect to={{ pathname: '/check-email', state: { userEmail } }} />;
-  }
 
   return (
     <ForgotPasswordForm