+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Correctly display session-remaining time after refresh. Refs STCOR-1087.
 * Rename `button.new` from "+ New" to "New". Refs STCOR-1093.
 * Supply `rtrIgnore` on forgot-password and forgot-username API requests. Refs STCOR-1099.
+* Codesmell cleanup in ForgotUserName, ForgotPassword. Refs STCOR-993.
 
 ## [11.1.1](https://github.com/folio-org/stripes-core/tree/v11.1.1) (2026-04-20)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v11.1.0...v11.1.1)